WEA-28 Cast temperature to integer in realm to entity mapper instead of at entity model

+++ app/src/main/java/com/ayush/weatherapp/entities/forecast/DailyDataEntity.java
@@ -83,16 +83,16 @@ public class DailyDataEntity implements Parcelable {
     return temperatureHigh;
   }
 
-  public void setTemperatureHigh(double temperatureHigh) {
-    this.temperatureHigh = (int) temperatureHigh;
+  public void setTemperatureHigh(int temperatureHigh) {
+    this.temperatureHigh = temperatureHigh;
   }
 
   public int getTemperatureLow() {
     return temperatureLow;
   }
 
-  public void setTemperatureLow(double temperatureLow) {
-    this.temperatureLow = (int) temperatureLow;
+  public void setTemperatureLow(int temperatureLow) {
+    this.temperatureLow = temperatureLow;
   }
 
   public double getWindSpeed() {
